requirements.txt
fastapi[all]
sqlalchemy
asyncpg
typing
python-dotenv
alembic
fastapi-cache2[redis]
celery
flower